release-changelog: Improve discovery of previous release tag
When we create multiple tags for a single commit, then git describe
--tags will only return one of them (probably the latest one?)[1] and as
such the discovery of the actual release tag may fail. We can workaround
that by simply returning all the tags that match a specific commit and
then perform the first-level filtering there.

Due to the pipeline depth we need to disable some checks from shellcheck
in order to keep things working.

[1] https://stackoverflow.com/questions/8089002/git-describe-with-two-tags-on-the-same-commit